# AUTOMATION/web/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulglade — API du studio",
    "description": "\nBackend local du studio Soulglade. Aucune authentification : le serveur\nn'accepte que ce qui vient de lui-même (voir `api/security.py`).\n\n**Conventions transverses**\n\n* presque toute route lit `?character=<id>` et le valide avant de toucher au\n  disque ; `/img` l'exige (pas de repli sur un personnage par défaut) ;\n* toute réponse porte un corps JSON, succès comme échec. Une erreur a la forme\n  `{\"ok\": false, \"erreur\": \"<texte français destiné à l'écran>\"}` ;\n* les uploads voyagent en base64 dans un corps JSON, jamais en\n  `multipart/form-data` — conséquence voulue du garde d'origine.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/app/stop": {
      "post": {
        "summary": "Arrêter le tableau de bord",
        "description": "Stops THIS web server. Answers first, exits after — otherwise the\nbrowser never receives the confirmation.",
        "operationId": "stop_dashboard_api_app_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/restart": {
      "post": {
        "summary": "Redémarrer le tableau de bord",
        "description": "Restarts THIS server (os.execv): same process ID, same window, code and\nconfig re-read cold. A real restart, not a data reload — the only way to\npick up a code change without going back to run_web.bat by hand.\n\nSurvives the move to uvicorn because web/app.py is still the entry point:\n`sys.argv` is its own, so re-executing it relaunches the same command line.",
        "operationId": "restart_dashboard_api_app_restart_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/comfy/stop": {
      "post": {
        "summary": "Arrêter ComfyUI",
        "operationId": "stop_comfy_api_app_comfy_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "ComfyUI ne tournait pas"
          }
        }
      }
    },
    "/api/app/comfy/stats": {
      "get": {
        "summary": "RAM / VRAM / thermique",
        "description": "Machine memory and thermals, for the banner and the Application screen.\n\nNo character parameter: VRAM/RAM belong to the MACHINE, one ComfyUI for\nthe whole platform (2026-09-01) — there is nothing here a character could\nscope.\n\nBoth probes go into a THREAD: `comfy_alive` cost a 2005 ms event-loop\nfreeze on 24/08 for having probed while blocking, and we are not replaying\nthat. The result is kept for a second and a half.\n\nTWO TRAPS, both observed on 30/08 with ComfyUI stopped — the case where the\nprobe is SLOW (urlopen waiting on a dead port, ~2 s):\n\n  1. The timestamp is set AFTER the work, not before. Stamped at the start\n     of the request, a probe longer than the TTL made the cache stale at\n     birth: it NEVER served anything again, and every call re-ran\n     nvidia-smi and the wait. Measured: 2087 ms on a call meant to come out\n     of the cache.\n  2. A lock, otherwise two concurrent calls both do the work. The second\n     waits for the first and reads its result.",
        "operationId": "get_comfy_stats_api_app_comfy_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComfyStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/comfy/unload": {
      "post": {
        "summary": "Décharger la VRAM",
        "description": "Unloads models and VRAM. Explicit gesture, never automatic.\n\nRefused during a batch: unloading under a running job would make it fail,\nand the user would lose a production to gain some VRAM. No character\nparameter, same reason as `get_comfy_stats` above.",
        "operationId": "unload_comfy_api_app_comfy_unload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Une production est en cours"
          },
          "502": {
            "description": "Le déchargement a échoué"
          }
        }
      }
    },
    "/api/app/comfy/restart": {
      "post": {
        "summary": "Redémarrer ComfyUI",
        "description": "Stops then restarts ComfyUI. Fire-and-forget: the recovery already shows\non the header's green dot (it polls /api/state in a loop), no need for one\nmore dedicated state to maintain.",
        "operationId": "restart_comfy_api_app_comfy_restart_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "erreur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erreur"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ActionResponse",
        "description": "The universal answer of an action route, `{ok: bool, erreur?: str}`.\n\nAUDIT §5.4: the convention holds on 4xx and 5xx too — `api.js` reads a JSON\nbody on every response, whatever the status.\n\nALWAYS DECLARE IT WITH `response_model_exclude_unset=True`. The old handlers\nreturned `{\"ok\": true}` and nothing else on success; a plain response_model\nwould add `\"erreur\": null` to every one of them. Harmless to `erreurDe()`,\nwhich only tests `ok === false` — but this migration is meant to keep the\nbodies identical, and a shape that drifts once drifts twice."
      },
...
      "ComfyStatsResponse": {
        "properties": {
          "en_ligne": {
            "type": "boolean",
            "title": "En Ligne"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "ram": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "vram": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VramUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "gpu": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Gpu"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "en_ligne"
        ],
        "title": "ComfyStatsResponse",
        "description": "RAM / VRAM / thermals. Always a dict, never an exception: it feeds a\ndisplay, not a decision. `en_ligne` says whether ComfyUI answered."
      },
...
      "ErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ErrorResponse",
        "description": "What every rejection looks like. `erreur` is French: it is displayed to\nthe user verbatim (AUDIT §5.5)."
      },
...
      "MemoryUsage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "libre": {
            "type": "integer",
            "title": "Libre"
          },
          "utilisee": {
            "type": "integer",
            "title": "Utilisee"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "total",
          "libre",
          "utilisee"
        ],
        "title": "MemoryUsage"
      },
...
      "VramUsage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "libre": {
            "type": "integer",
            "title": "Libre"
          },
          "utilisee": {
            "type": "integer",
            "title": "Utilisee"
          },
          "nom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nom"
          },
          "torch_reserve": {
            "type": "integer",
            "title": "Torch Reserve",
            "default": 0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "total",
          "libre",
          "utilisee"
        ],
        "title": "VramUsage"
      },
...
    }
  }
}